requirements.txt
pyrogram==1.4.16
TgCrypto
py-tgcalls==0.9.7
pyyaml==5.4
python-dotenv
youtube_search_python
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg-python
ffmpeg
Pillow
ujson
rich
async-lru
pyaes
PySocks
motor
dnspython
speedtest-cli
wget
apscheduler
yt_dlp
cachetools==5.1.0
git+https://github.com/ADITYAxSERVER/downloader
numpy
pykeyboard
